qr-code-reader/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pillow==11.0.0
pyzbar==0.1.9
requests==2.32.3
urllib3==2.2.3
